requirements-dev.txt
anyio==3.6.2 ; python_version >= "3.8" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.8" and python_version < "4.0"
black==22.12.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.3.0 ; python_version >= "3.8" and python_version < "4.0"
h11==0.12.0 ; python_version >= "3.8" and python_version < "4.0"
httpcore==0.14.7 ; python_version >= "3.8" and python_version < "4.0"
httpx==0.22.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
jianshuresearchtools==2.10.1 ; python_version >= "3.8" and python_version < "4.0"
jieba==0.42.1 ; python_version >= "3.8" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
lxml==4.8.0 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
nodeenv==1.7.0 ; python_version >= "3.8" and python_version < "4.0"
pathspec==0.11.0 ; python_version >= "3.8" and python_version < "4.0"
platformdirs==2.6.2 ; python_version >= "3.8" and python_version < "4.0"
prettytable==3.6.0 ; python_version >= "3.8" and python_version < "4.0"
pyecharts==2.0.1 ; python_version >= "3.8" and python_version < "4.0"
pymongo==4.3.3 ; python_version >= "3.8" and python_version < "4.0"
pyright==1.1.292 ; python_version >= "3.8" and python_version < "4.0"
pywebio==1.7.1 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "4.0"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.8" and python_version < "4.0"
ruff==0.0.241 ; python_version >= "3.8" and python_version < "4.0"
setuptools==67.1.0 ; python_version >= "3.8" and python_version < "4.0"
simplejson==3.18.2 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.8" and python_full_version < "3.11.0a7"
tornado==6.2 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "3.10"
ua-parser==0.16.1 ; python_version >= "3.8" and python_version < "4.0"
user-agents==2.2.0 ; python_version >= "3.8" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.8" and python_version < "4.0"
